BUG/MEDIUM: kqueue: Make sure we report read events even when no data.
When we have a EVFILT_READ event, an optimization was made, and the FD was
not reported as ready to receive if there were no data available. That way,
if the socket was closed by our peer (the EV8EOF flag was set), and there were
no remaining data to read, we would just close(), and avoid doing a recv().
However, it may be fine for TCP socket, but it is not for UDP.
If we send data via UDP, and we receive an error, the only way to detect it
is to attempt a recv(). However, in this case, kevent() will report a read
event, but with no data, so we'd just ignore that read event, nothing would be
done about it, and the poller would be woken up by it over and over.
To fix this, report read events if either we have data, or the EV_EOF flag
is not set.

This should be backported to 2.1, 2.0, 1.9 and 1.8.

(cherry picked from commit eaefc3c5032506e89cceb6ad5fdd1c5955c4ea66)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 66c7b64e20c3b9e93675fee50f245e5257e393ea)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 7b869bf019638ae4dcb5e312f7134ed485aba5b6)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 674bfb73450c41fe59e0ec3312d34e1d80b4f6e0)
Signed-off-by: Willy Tarreau <w@1wt.eu>