BUG/MINOR: tasks: only requeue a task if it was already in the queue
Commit 0742c314c3 ("BUG/MEDIUM: tasks: Make sure we switch wait queues
in task_set_affinity().") had a slight side effect on expired timeouts,
which is that when used before a timeout is updated, it will cause an
existing task to be requeued earlier than its expected timeout when done
before being updated, resulting in the next poll wakup timeout too early
or even instantly if the previous wake up was done on a timeout. This is
visible in strace when health checks are enabled because there are two
poll calls, one of which has a short or zero delay. The correct solution
is to only requeue a task if it was already in the queue.

This can be backported to all branches having the fix above.

(cherry picked from commit 440d09b2448fcddea6877054300a95ba5b55dac7)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 5f9b548c65dceff150154ed1152b35e4ff94dc1f)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit fcd7401b3e694a43f94744776831a4eaccdfdfe6)
Signed-off-by: Willy Tarreau <w@1wt.eu>