BUG/MINOR: sample: fix the closing bracket and LF in the debug converter
The closing bracket was emitted for the "debug" converter even when the
opening one was not sent, and the new line was not always emitted. Let's
fix this. This is harmless since this converter is not built by default.

(cherry picked from commit 50603267981a002d2593bfe219e5071d66a8ea65)
Signed-off-by: Willy Tarreau <w@1wt.eu>