BUG/MINOR: sample: always check converters' arguments
In 1.5-dev20, sample-fetch arguments parsing was addresse by commit
689a1df0a1 ("BUG/MEDIUM: sample: simplify and fix the argument parsing").
The issue was that argument checks were not run for sample-fetches if
parenthesis were not present. Surprisingly, the fix was mde only for
sample-fetches and not for converters which suffer from the exact same
problem. There are even a few comments in the code mentioning that some
argument validation functions are not called when arguments are missing.

This fix applies the exact same method as the one above. The impact of
this bug is limited because over the years the code has learned to work
around this issue instead of fixing it.

This may be backported to all maintained versions.

(cherry picked from commit 46dfd78cbf2591ab9cd2d8e30046e811f4179f7b)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 4a25605ab4475130b56e9b25d3ce1d071b796e2a)
Signed-off-by: Willy Tarreau <w@1wt.eu>