BUG/MEDIUM: state-file: do not allocate a full buffer for each server entry
Starting haproxy with a state file of 700k servers eats 11.2 GB of RAM
due to a mistake in the function that loads the strings into a tree: it
allocates a full buffer for each backend+server name instead of allocating
just the required string. By just fixing this we're down to 80 MB.

This should be backported to 2.1.

(cherry picked from commit 7d6a1fa311312bb99b98f548399fc30fc7802ad7)
Signed-off-by: Willy Tarreau <w@1wt.eu>